feat: 혼잡도 타입 변경
int -> double 로 변경해야 정규화한 의미가 있음

+++ ai_model/src/main/java/com/capstone/ai_model/dto/TrainingData.java
@@ -7,7 +7,7 @@ import lombok.ToString;
 @AllArgsConstructor
 public class TrainingData {
     double[] morningFeature;
-    int morningCongestion;
+    double morningCongestion;
     double[] eveningFeature;
-    int eveningCongestion;
+    double eveningCongestion;
 }
